refactor(vrli): restructured code Signed-off-by: Stefan Genov <stefan.genov@broadcom.com>
