+++ common/artifact-manager/src/main/java/com/vmware/pscoe/iac/artifact/IMPORTANT.md
@@ -6,7 +6,7 @@ The structure of this folder is being reworked. Classes are being moved based on
 
 - [x] Aria Automation
 - [ ] Aria Orchestrator
-- [ ] Aria Logs
+- [x] Aria Logs
 - [x] Aria Operations
 - [ ] SSH
 - [ ] vCD
