+++ common/artifact-manager/src/main/java/com/vmware/pscoe/iac/artifact/PackageStoreFactory.java
@@ -19,13 +19,16 @@ import java.util.List;
 
 import com.vmware.pscoe.iac.artifact.configuration.*;
 import com.vmware.pscoe.iac.artifact.rest.*;
-import com.vmware.pscoe.iac.artifact.rest.client.vrli.RestClientVrliV1;
-import com.vmware.pscoe.iac.artifact.rest.client.vrli.RestClientVrliV2;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.vmware.pscoe.iac.artifact.aria.automation.rest.RestClientVraNg;
+import com.vmware.pscoe.iac.artifact.aria.logs.configuration.ConfigurationVrli;
+import com.vmware.pscoe.iac.artifact.aria.logs.rest.v1.RestClientVrliV1;
+import com.vmware.pscoe.iac.artifact.aria.logs.rest.v2.RestClientVrliV2;
+import com.vmware.pscoe.iac.artifact.aria.logs.store.v1.VrliPackageStoreV1;
+import com.vmware.pscoe.iac.artifact.aria.logs.store.v2.VrliPackageStoreV2;
 import com.vmware.pscoe.iac.artifact.aria.operations.configuration.ConfigurationVrops;
 import com.vmware.pscoe.iac.artifact.aria.operations.rest.RestClientVrops;
 import com.vmware.pscoe.iac.artifact.aria.operations.store.VropsPackageStore;
