+++ common/artifact-manager/src/main/java/com/vmware/pscoe/iac/artifact/aria/logs/rest/v1/models/ContentPackDTO.java
@@ -1,0 +1,177 @@
+
+/*
+ * #%L
+ * artifact-manager
+ * %%
+ * Copyright (C) 2023 VMware
+ * %%
+ * Build Tools for VMware Aria
+ * Copyright 2023 VMware, Inc.
+ * 
+ * This product is licensed to you under the BSD-2 license (the "License"). You may not use this product except in compliance with the BSD-2 License.  
+ * 
+ * This product may include a number of subcomponents with separate copyright notices and license terms. Your use of these subcomponents is subject to the terms and conditions of the subcomponent's license, as noted in the LICENSE file.
+ * #L%
+ */
+package com.vmware.pscoe.iac.artifact.aria.logs.rest.v1.models;
+
+import java.io.Serializable;
+import java.util.HashMap;
+import java.util.Map;
+
+import com.fasterxml.jackson.annotation.JsonAnyGetter;
+import com.fasterxml.jackson.annotation.JsonAnySetter;
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.annotation.JsonPropertyOrder;
+
+@JsonInclude(JsonInclude.Include.NON_NULL)
+@JsonPropertyOrder({ "name", "namespace", "formatVersion", "contentVersion", "author", "url", "info", "instructions",
+		"upgradeInstructions", "framework" })
+public class ContentPackDTO implements Serializable {
+	private final static long serialVersionUID = 7183649262190355102L;
+
+	@JsonProperty("name")
+	private String name;
+
+	@JsonProperty("namespace")
+	private String namespace;
+
+	@JsonProperty("formatVersion")
+	private String formatVersion;
+
+	@JsonProperty("contentVersion")
+	private String contentVersion;
+
+	@JsonProperty("author")
+	private String author;
+
+	@JsonProperty("url")
+	private String url;
+
+	@JsonProperty("info")
+	private String info;
+
+	@JsonProperty("instructions")
+	private String instructions;
+
+	@JsonProperty("upgradeInstructions")
+	private String upgradeInstructions;
+
+	@JsonProperty("framework")
+	private String framework;
+
+	@JsonIgnore
+	private transient Map<String, Object> additionalProperties = new HashMap<>();
+
+	@JsonProperty("name")
+	public String getName() {
+		return name;
+	}
+
+	@JsonProperty("name")
+	public void setName(String name) {
+		this.name = name;
+	}
+
+	@JsonProperty("namespace")
+	public String getNamespace() {
+		return namespace;
+	}
+
+	@JsonProperty("namespace")
+	public void setNamespace(String namespace) {
+		this.namespace = namespace;
+	}
+
+	@JsonProperty("formatVersion")
+	public String getFormatVersion() {
+		return formatVersion;
+	}
+
+	@JsonProperty("formatVersion")
+	public void setFormatVersion(String formatVersion) {
+		this.formatVersion = formatVersion;
+	}
+
+	@JsonProperty("contentVersion")
+	public String getContentVersion() {
+		return contentVersion;
+	}
+
+	@JsonProperty("contentVersion")
+	public void setContentVersion(String contentVersion) {
+		this.contentVersion = contentVersion;
+	}
+
+	@JsonProperty("author")
+	public String getAuthor() {
+		return author;
+	}
+
+	@JsonProperty("author")
+	public void setAuthor(String author) {
+		this.author = author;
+	}
+
+	@JsonProperty("url")
+	public String getUrl() {
+		return url;
+	}
+
+	@JsonProperty("url")
+	public void setUrl(String url) {
+		this.url = url;
+	}
+
+	@JsonProperty("info")
+	public String getInfo() {
+		return info;
+	}
+
+	@JsonProperty("info")
+	public void setInfo(String info) {
+		this.info = info;
+	}
+
+	@JsonProperty("instructions")
+	public String getInstructions() {
+		return instructions;
+	}
+
+	@JsonProperty("instructions")
+	public void setInstructions(String instructions) {
+		this.instructions = instructions;
+	}
+
+	@JsonProperty("upgradeInstructions")
+	public String getUpgradeInstructions() {
+		return upgradeInstructions;
+	}
+
+	@JsonProperty("upgradeInstructions")
+	public void setUpgradeInstructions(String upgradeInstructions) {
+		this.upgradeInstructions = upgradeInstructions;
+	}
+
+	@JsonProperty("framework")
+	public String getFramework() {
+		return framework;
+	}
+
+	@JsonProperty("framework")
+	public void setFramework(String framework) {
+		this.framework = framework;
+	}
+
+	@JsonAnyGetter
+	public Map<String, Object> getAdditionalProperties() {
+		return this.additionalProperties;
+	}
+
+	@JsonAnySetter
+	public void setAdditionalProperties(String name, Object value) {
+		this.additionalProperties.put(name, value);
+	}
+}
