--- common/artifact-manager/src/main/java/com/vmware/pscoe/iac/artifact/rest/client/messages/Errors.java
@@ -1,22 +1,0 @@
-/*-
- * #%L
- * artifact-manager
- * %%
- * Copyright (C) 2023 VMware
- * %%
- * Build Tools for VMware Aria
- * Copyright 2023 VMware, Inc.
- * 
- * This product is licensed to you under the BSD-2 license (the "License"). You may not use this product except in compliance with the BSD-2 License.  
- * 
- * This product may include a number of subcomponents with separate copyright notices and license terms. Your use of these subcomponents is subject to the terms and conditions of the subcomponent's license, as noted in the LICENSE file.
- * #L%
- */
-package com.vmware.pscoe.iac.artifact.rest.client.messages;
-
-public class Errors {
-	public static final String MAP_ALERT_DATA_ERROR = "Unable to map alert data";
-	public static final String PROCESS_ALERT_DATA_ERROR = "Unable to process alert data";
-	public static final String MAP_CONTENT_PACK_DATA_ERROR = "Unable to map content pack data";
-	public static final String PROCESS_CONTENT_PACK_DATA_ERROR = "Unable to process content pack data";
-}
