--- common/artifact-manager/src/main/java/com/vmware/pscoe/iac/artifact/rest/model/vrli/v2/package-info.java
@@ -1,20 +1,0 @@
-/*-
- * #%L
- * artifact-manager
- * %%
- * Copyright (C) 2023 VMware
- * %%
- * Build Tools for VMware Aria
- * Copyright 2023 VMware, Inc.
- * 
- * This product is licensed to you under the BSD-2 license (the "License"). You may not use this product except in compliance with the BSD-2 License.
- * 
- * This product may include a number of subcomponents with separate copyright notices and license terms. Your use of these subcomponents is subject to the terms and conditions of the subcomponent's license, as noted in the LICENSE file.
- * #L%
- */
-
-/**
- * Package that represents vRLI API v2 Rest Model.
- *
- */
-package com.vmware.pscoe.iac.artifact.rest.model.vrli.v2;
