+++ common/artifact-manager/src/test/java/com/vmware/pscoe/iac/artifact/aria/.gitignore
@@ -1,0 +1,2 @@
+# `logs` are excluded generally, so, re-include them
+!logs
